feat: add authorization to query handlers & reorder permission checks
- Move require_permission after entity existence checks in all
  organization command handlers to prevent information leakage
- Add caller_id/org_id and permission enforcement to query handlers
  across organization, workspace, project, and task bounded contexts
- Add organization existence validation to handlers that lacked it
  (add_org_storage, add_sso_integration, create_department,
   create_org_role, create_team)
- Add domain exceptions for organization, department, org_membership,
  org_role; extend team_exceptions
- Add HTTP 409 status and error codes to DuplicateInvitationForEmail
  and MemberAlreadyExists exceptions
- Add Project and Task BC presentation layer (controllers, schemas,
  dependencies, API routes)
- Add workspace query authorization and new workspace endpoints
- Add migration for storage_quota BigInteger column
- Update E2E test infrastructure and add scenario tests
- Add .windsurf/ to .gitignore

+++ app/context/organization/infrastructure/persistence/orm_models/storage_integration_orm.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import uuid
 
-from sqlalchemy import ForeignKey, Integer, JSON, String, Text
+from sqlalchemy import BigInteger, ForeignKey, Integer, JSON, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.shared.infrastructure.persistence.sqlalchemy_base_orm_model import BaseORMModel
@@ -26,7 +26,7 @@ class StorageIntegrationORM(BaseORMModel):
     sc_access_key: Mapped[str] = mapped_column(Text, nullable=False, default="")
 
     # --- Embedded StorageQuota VO ---
-    sq_max_bytes: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    sq_used_bytes: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    sq_max_file_size_bytes: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    sq_max_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
+    sq_used_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
+    sq_max_file_size_bytes: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     sq_allowed_extensions: Mapped[list | None] = mapped_column(JSON, nullable=True)
